use i64 in bucket

+++ src/common/connection.rs
@@ -6,7 +6,7 @@ use uuid::Uuid;
 pub struct Connection {
     pub device_id: Uuid,
     pub ip: Option<String>,
-    pub bucket: Option<i32>,
+    pub bucket: Option<i64>,
     pub last_connection: DateTime<Utc>,
     pub next_wakeup: Option<DateTime<Utc>>,
     pub status: ConnectionStatus,
@@ -16,7 +16,7 @@ impl Connection {
     pub fn new(
         device_id: Uuid,
         ip: Option<String>,
-        bucket: Option<i32>,
+        bucket: Option<i64>,
         status: ConnectionStatus,
     ) -> Self {
         Connection {
